backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
dj-database-url>=0.5.0
django-cors-headers>=3.13.0
psycopg2-binary>=2.9.3
gunicorn>=20.1.0